[backend] Fix cyclic dependency
