+++ opencti-platform/opencti-graphql/src/http/httpPlatform.js
@@ -15,12 +15,12 @@ import rateLimit from 'express-rate-limit';
 import contentDisposition from 'content-disposition';
 import { basePath, booleanConf, DEV_MODE, ENABLED_UI, logApp, OPENCTI_SESSION } from '../config/conf';
 import passport, { isStrategyActivated, STRATEGY_CERT } from '../config/providers';
-import { HEADERS_AUTHENTICATORS, loginFromProvider, sessionAuthenticateUser, userWithOrigin } from '../domain/user';
+import { authenticateUserFromRequest, HEADERS_AUTHENTICATORS, loginFromProvider, sessionAuthenticateUser, userWithOrigin } from '../domain/user';
 import { downloadFile, getFileContent, isStorageAlive, loadFile } from '../database/file-storage';
 import createSseMiddleware from '../graphql/sseMiddleware';
 import initTaxiiApi from './httpTaxii';
 import initHttpRollingFeeds from './httpRollingFeed';
-import { DEFAULT_INVALID_CONF_VALUE, executionContext, SYSTEM_USER } from '../utils/access';
+import { DEFAULT_INVALID_CONF_VALUE, executionContext, isBypassUser, SYSTEM_USER } from '../utils/access';
 import { ENTITY_TYPE_SETTINGS } from '../schema/internalObject';
 import { getEntityFromCache } from '../database/cache';
 import { isEmptyField, isNotEmptyField } from '../database/utils';
@@ -29,7 +29,43 @@ import { internalLoadById } from '../database/middleware-loader';
 import { delUserContext, redisIsAlive } from '../database/redis';
 import { rabbitMQIsAlive } from '../database/rabbitmq';
 import { isEngineAlive } from '../database/engine';
-import { createContext } from './httpServer';
+
+export const createContext = async (req, res, contextName) => {
+  const executeContext = executionContext(contextName);
+  executeContext.req = req;
+  executeContext.res = res;
+  const settings = await getEntityFromCache(executeContext, SYSTEM_USER, ENTITY_TYPE_SETTINGS);
+  executeContext.otp_mandatory = settings.otp_mandatory ?? false;
+  executeContext.workId = req.headers['opencti-work-id']; // Api call comes from a worker processing
+  executeContext.draft_context = req.headers['opencti-draft-id']; // Api call is to be made is specific draft context
+  executeContext.eventId = req.headers['opencti-event-id']; // Api call is due to listening event
+  executeContext.previousStandard = req.headers['previous-standard']; // Previous standard id
+  executeContext.synchronizedUpsert = req.headers['synchronized-upsert'] === 'true'; // If full sync needs to be done
+  try {
+    const user = await authenticateUserFromRequest(executeContext, req);
+    if (user) {
+      if (!Object.keys(req.headers).some((k) => k === 'opencti-draft-id')) {
+        executeContext.draft_context = user.draft_context;
+      }
+      executeContext.user = userWithOrigin(req, user);
+      executeContext.user_otp_validated = true;
+      executeContext.user_with_session = isNotEmptyField(req.session?.user);
+      if (executeContext.user_with_session) {
+        executeContext.user_otp_validated = req.session?.user.otp_validated ?? false;
+      }
+      if (isBypassUser(executeContext.user)) {
+        executeContext.user_inside_platform_organization = true;
+      } else {
+        const userOrganizationIds = (user.organizations ?? []).map((organization) => organization.internal_id);
+        executeContext.user_inside_platform_organization = settings.platform_organization
+          ? userOrganizationIds.includes(settings.platform_organization) : true;
+      }
+    }
+  } catch (error) {
+    logApp.error('Fail to authenticate the user in graphql context hook', { cause: error });
+  }
+  return executeContext;
+};
 
 const setCookieError = (res, message) => {
   res.cookie('opencti_flash', message || 'Unknown error', {
